Update Index.cshtml
Using ASP.net, I implemented a simple web app that can process C# code and write values to a dynamic site. At the moment the site is mostly templated and outputs the smallest factorial after 1e6. In the future we will connect the existing React framework to this backend, and eventually also the python machine learning code (using a web form).